Move worklist type toggle out of filters
