+++ src/js/views/patients/worklist/layout.hbs
@@ -1,5 +1,8 @@
 <div class="list-page__header">
-  <div class="list-page__title"><span data-title-region></span><span data-tooltip-region></span></div>
+  <div class="flex list-page__title">
+    <div data-title-region></div>
+    <div class="worklist-list__toggle" data-toggle-region></div>
+  </div>
   <div class="list-page__filters flex">
     <div class="worklist-list__select-all" data-select-all-region></div>
     <div class="worklist-list__filter-region" data-filters-region></div>
