+++ src/js/views/patients/worklist/worklist-list.scss
@@ -7,6 +7,10 @@
   }
 }
 
+.worklist__type-toggle {
+  margin-left: auto;
+}
+
 .worklist__header {
   table-layout: fixed;
 }
